Move Graphics to top-level type; use idiomatic Swift style to load a Font
